+++ CHANGELOG.md
@@ -6,11 +6,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+* `replaceCargoLockHook` can now be used to easily replace or insert a
+  `Cargo.lock` file in the current derivation
+
 ### Changed
 * `cargoAudit` will pass `--ignore yanked` by default if `cargoAuditExtraArgs`
   are not specified. This is because `cargo-audit` cannot check for yanked
   crates from inside of the sandbox. To get the old behavior back, set
   `cargoAuditExtraArgs = "";`.
+* `mkCargoDerivation` (and by extension anything which delegates to it) will now
+  automatically use the value of `cargoLock` or the contents of
+  `cargoLockContents`/`cargoLockParsed` to replace the workspace `Cargo.lock`
+  file. To disable this behavior, set `doNotReplaceCargoLock = true;`.
 
 ### Fixed
 * Fixed handling of Cargo workspace inheritance for git-dependencies where said